requirements.txt
# Discord Bot Requirements

# Core Discord library
discord.py>=2.3.0

# Web framework
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
jinja2>=3.1.0
python-multipart>=0.0.6

# Database
sqlalchemy[asyncio]>=2.0.0
aiosqlite>=0.19.0  # For SQLite async support
asyncpg>=0.29.0    # For PostgreSQL async support (optional)

# HTTP client for crawler
aiohttp>=3.9.0

# Environment variables
python-dotenv>=1.0.0

# Caching (optional)
redis>=5.0.0
aioredis>=2.0.0

# Data processing
feedparser>=6.0.0  # For RSS feed parsing
beautifulsoup4>=4.12.0  # For HTML parsing
lxml>=4.9.0  # XML parser

# Utilities
python-dateutil>=2.8.0
pytz>=2023.3

# Logging
coloredlogs>=15.0  # For colored console logs (optional)

# Development dependencies (uncomment for development)
# pytest>=7.4.0
# pytest-asyncio>=0.21.0
# black>=23.0.0
# flake8>=6.0.0
# mypy>=1.5.0

# Optional: Image processing
# Pillow>=10.0.0

# Optional: Data analysis
# pandas>=2.0.0
# numpy>=1.24.0

# Optional: Machine learning
# scikit-learn>=1.3.0
# torch>=2.0.0  # If using PyTorch

# Optional: Monitoring
# prometheus-client>=0.17.0